ViewController#countDisksをBoardに移し、そちらを使用するよう修正。テストも併せて作成
