+++ ReversiTests/Model/ReversiGameTest.swift
@@ -19,6 +19,8 @@ final class ReversiGameTest: XCTestCase {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
 
+    // MARK: setDisk, diskAt, countDisks
+
     func test_白のディスクをボードの一番先頭に置ける() throws {
         XCTAssertEqual(game.board.countDisks(of: .light), 0)
 
@@ -35,5 +37,25 @@ final class ReversiGameTest: XCTestCase {
 
         XCTAssertEqual(game.board.countDisks(of: .dark), 1)
         XCTAssertEqual(game.board.diskAt(x: 7, y: 7), .dark)
+    }
+
+    // MARK: sideWithMoreDisks
+
+    func test_黒のディスクが1つ置かれているとディスクの枚数が多いのは黒() throws {
+        game.board.setDisk(.dark, atX: 0, y: 0)
+
+        XCTAssertEqual(game.board.sideWithMoreDisks(), .dark)
+    }
+
+    func test_白のディスクが1つ置かれているとディスクの枚数が多いのは白() throws {
+        game.board.setDisk(.light, atX: 0, y: 0)
+
+        XCTAssertEqual(game.board.sideWithMoreDisks(), .light)
+    }
+
+    func test_左半分に黒_右半分に白のディスクが置かれていると引き分け() throws {
+        let tiedGame = tiedComputerMatchWithLeftSideDarkAndRightSideLightBoard()
+
+        XCTAssertNil(tiedGame.board.sideWithMoreDisks())
     }
 }
