+++ ReversiTests/Repository/ReversiGameRepositoryTest.swift
@@ -79,29 +79,6 @@ final class ReversiGameRepositoryTest: XCTestCase {
 
         XCTAssertEqual(actual, expected)
     }
-
-    // MARK: Helper
-
-    private func newGameStartFromLight() -> ReversiGame {
-        var game = ReversiGame.newGame()
-        game.turn?.flip()
-        return game
-    }
-
-    private func tiedComputerMatchWithLeftSideDarkAndRightSideLightBoard() -> ReversiGame {
-        var game = ReversiGame.newGame()
-        game.turn = nil
-        game.playerControls = [.computer, .computer]
-        for y in game.board.yRange {
-            for x in 0..<game.board.height / 2 {
-                game.board.setDisk(.dark, atX: x, y: y)
-            }
-            for x in game.board.height / 2..<game.board.height {
-                game.board.setDisk(.light, atX: x, y: y)
-            }
-        }
-        return game
-    }
 }
 
 private class FakeFileSaveAndLoadStrategy: FileSaveAndLoadStrategy {
